requirements.txt
alembic==1.0.0
blinker==1.4
certifi==2018.8.24
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Dance==1.0.0
Flask-Env==2.0.0
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
lazy==1.3
Mako==1.0.7
MarkupSafe==1.0
oauthlib==2.1.0
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
SQLAlchemy==1.2.11
SQLAlchemy-Utils==0.33.3
urllib3==1.23
URLObject==2.4.3
Werkzeug==0.14.1
alembic==1.0.0
blinker==1.4
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Dance==1.0.0
Flask-Env==2.0.0
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-OAuth==0.12
Flask-SQLAlchemy==2.3.2
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
gunicorn==19.9.0
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
lazy==1.3
Mako==1.0.7
MarkupSafe==1.0
oauth2==1.9.0.post1
oauthlib==2.1.0
psycopg2==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.7
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
six==1.11.0
SQLAlchemy==1.2.11
SQLAlchemy-Utils==0.33.3
uritemplate==3.0.0
urllib3==1.23
URLObject==2.4.3
Werkzeug==0.14.1
